illustration_agent/requirements.txt
google-adk==1.8.0
a2a-sdk==0.2.16